requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# pip install -r requirements.txt
# pip3 install --break-system-packages -r requirements.txt
#

discord==2.3.2
python-dotenv==1.0.1
audioop-lts; python_version>='3.13'
